SearchBox를 SearchBar 밖으로 이동
